Controlable Camera
I have enabled the camera to easily be moved around a scene if set as the player object.

I plan to create three scenes for the engine next, this will act as the three 'base' scenes, they will be simple but also large enough for a user to place other game objects inside.

Following the creaton of the scenes I will need to find a way to load them into the game engine from a JSON file.