Movement component
I have created a movement component which can be connected to a game object and will enable it to be moved with some movement keys.

I have also begun work on a rendering script with the intent to allow for a game object to be passed into it and it will be drawn.