Render Object
I have created a class which when a game object is passed into it it will be drawn into the world.

Next I aim to add a method to create different camera types: static (doesn't move), first-person (acts as a head on the object it is attached to) and third-person (follows the player at a slight angle).